Migrate encoder FieldType methods into subpackage
