+++ supyr_struct/field_type_methods.py
@@ -33,8 +33,8 @@ be working with the same parameter and return data types.
 
 from decimal import Decimal
 from math import ceil, log
-from struct import pack, pack_into, unpack
-from time import mktime, ctime, strptime
+from struct import pack_into, unpack
+from time import ctime
 
 from supyr_struct.defs.constants import NAME, UNNAMED, DEFAULT, NODE_CLS,\
      COMPUTE_READ, COMPUTE_WRITE, COMPUTE_SIZECALC, STEPTREE, TYPE, VALUE_MAP,\
@@ -64,12 +64,9 @@ __all__ = [
     # Decoders
     'decode_numeric', 'decode_string', 'no_decode',
     'decode_big_int', 'decode_bit_int', 'decode_raw_string',
-    # Encoders
-    'encode_numeric', 'encode_string', 'no_encode',
-    'encode_big_int', 'encode_bit_int',
 
     # wrapper functions
-    'encoder_wrapper', 'decoder_wrapper',
+    'decoder_wrapper',
 
     # Specialized routines
 
@@ -85,9 +82,6 @@ __all__ = [
     # Decoders
     'decode_24bit_numeric', 'decode_decimal', 'decode_bit',
     'decode_timestamp', 'decode_string_hex',
-    # Encoders
-    'encode_24bit_numeric', 'encode_decimal', 'encode_bit', 'encode_raw_string',
-    'encode_int_timestamp', 'encode_float_timestamp', 'encode_string_hex',
 
     # Sanitizer functions
     'bool_sanitizer', 'enum_sanitizer', 'switch_sanitizer',
@@ -130,16 +124,6 @@ def decoder_wrapper(de):
             self, rawdata, desc, parent, attr_index))
 
     return wrapped_decoder
-
-
-def encoder_wrapper(en):
-    '''
-    '''
-    def wrapped_encoder(
-            self, node, parent=None, attr_index=None, _encode=en):
-        return _encode(self, node.data, parent, attr_index)
-
-    return wrapped_encoder
 
 
 def format_parse_error(e, **kwargs):
@@ -1845,162 +1829,6 @@ def decode_bit_int(self, rawdata, desc=None, parent=None, attr_index=None):
 
     return bitint
 
-
-# #################################################
-'''############  Encoder functions  ############'''
-# #################################################
-
-
-def encode_numeric(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python int into a bytes representation.
-    Encoding is done using struct.pack
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    return self.struct_packer(node)
-
-
-def encode_decimal(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python Decimal into a bytes representation.
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    raise NotImplementedError('Encoding Decimal objects is not supported yet.')
-
-
-def encode_24bit_numeric(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python int to a signed or unsigned 24-bit bytes representation.
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    if self.enc[1] == 't':
-        # int can be signed
-        assert node >= -0x800000 and node <= 0x7fffff, (
-            '%s is too large to pack as a 24bit signed int.' % node)
-        if node < 0:
-            # int IS signed
-            node += 0x1000000
-    else:
-        assert node >= 0 and node <= 0xffffff, (
-            '%s is too large to pack as a 24bit unsigned int.' % node)
-
-    # pack and return the int
-    if self.endian == '<':
-        return pack('<I', node)[0:3]
-    return pack('>I', node)[1:4]
-
-
-def encode_int_timestamp(self, node, parent=None, attr_index=None):
-    '''
-    '''
-    return self.struct_packer(int(mktime(strptime(node))))
-
-
-def encode_float_timestamp(self, node, parent=None, attr_index=None):
-    '''
-    '''
-    return self.struct_packer(float(mktime(strptime(node))))
-
-
-def encode_string(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python string into a bytes representation,
-    making sure there is a delimiter character on the end.
-    Encoding is done using str.encode
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    if not node.endswith(self.str_delimiter):
-        return (node + self.str_delimiter).encode(self.enc)
-    return node.encode(self.enc)
-
-
-def encode_raw_string(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python string into a bytes representation.
-    Encoding is done using str.encode
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    return node.encode(self.enc)
-
-
-def encode_string_hex(self, node, parent=None, attr_index=None):
-    '''
-    Encodes a python string formatted as a hex string into a bytes object.
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    return int(node, 16).to_bytes((len(node) + 1)//2, 'big')
-
-
-def encode_big_int(self, node, parent=None, attr_index=None):
-    '''
-    Encodes arbitrarily sized signed or unsigned integers
-    on the byte level in either ones or twos compliment.
-    Encoding is done using int.to_bytes
-
-    Returns a bytes object encoded represention of the "node" argument.
-    '''
-    bytecount = parent.get_size(attr_index)
-
-    if not bytecount:
-        return b''
-
-    if self.endian == '<':
-        endian = 'little'
-    else:
-        endian = 'big'
-
-    if self.enc[-1] == 'S':
-        # twos compliment
-        return node.to_bytes(bytecount, endian, signed=True)
-    elif self.enc[-1] == 's':
-        # ones compliment
-        if node < 0:
-            return (node-1).to_bytes(bytecount, endian, signed=True)
-        return node.to_bytes(bytecount, endian, signed=False)
-
-    return node.to_bytes(bytecount, endian)
-
-
-def encode_bit(self, node, parent=None, attr_index=None):
-    '''
-    Encodes an int to a single bit.
-    Returns the encoded int, the offset it should be
-    shifted to, and a mask that covers its range.
-    '''
-    # return the int with the bit offset and a mask of 1
-    return(node, parent.ATTR_OFFS[attr_index], 1)
-
-
-def encode_bit_int(self, node, parent=None, attr_index=None):
-    '''
-    Encodes arbitrarily sized signed or unsigned integers
-    on the bit level in either ones or twos compliment
-
-    Returns the encoded int, the offset it should be
-    shifted to, and a mask that covers its range.
-    '''
-
-    bitcount = parent.get_size(attr_index)
-    offset = parent.ATTR_OFFS[attr_index]
-    mask = (1 << bitcount) - 1
-
-    # if the number is signed
-    if node < 0:
-        signmask = 1 << (bitcount - 1)
-        if self.enc == 'S':
-            # twos signed
-            return(2*signmask + node, offset, mask)
-        # ones signed
-        return(2*signmask + (node-1), offset, mask)
-    return(node, offset, mask)
-
-
 # ##########################################################
 '''#########  Void and Pad FieldType functions  #########'''
 # ##########################################################
@@ -2045,11 +1873,6 @@ def pad_serializer(self, node, parent=None, attr_index=None,
 def no_decode(self, rawdata, desc=None, parent=None, attr_index=None):
     ''''''
     return rawdata
-
-
-def no_encode(self, node, parent=None, attr_index=None):
-    ''''''
-    return node
 
 
 # ###################################################
